:ambulance: Update(listboard): Remove condition for displaying buttons
Removed a condition that was causing the endpoint buttons to only be displayed if the Coordinator Exit Modal object existed. This allows for the Care Seeking Endpoint button and the Cancer Dx Endpoint button to always be displayed on the listboard table button column regardless of whether or not the Coordinator Exit form has been filled. This improves the usability as it allows for navigation to these endpoints without needing to first complete the Coordinator Exit form.